+++ src/common/options.ts
@@ -1,12 +1,15 @@
-interface Option<T extends "boolean" | "string" | "number" = "boolean"> {
+import { map } from "./objects";
+
+export type OptionTypeString = "boolean" | "string" | "number";
+
+export type OptionTypeFromString<T extends OptionTypeString> =
+    T extends "boolean" ? boolean : T extends "string" ? string : number;
+
+export interface Option<T extends OptionTypeString> {
     name: string;
     description: string;
     type: T;
-    default: T extends "boolean"
-        ? boolean
-        : T extends "string"
-        ? string
-        : number;
+    default: OptionTypeFromString<T>;
 }
 
 const options = {
@@ -22,6 +25,29 @@ const options = {
         type: "boolean",
         default: false,
     },
-} as const satisfies Record<string, Option>;
+} as const satisfies Record<string, Option<OptionTypeString>>;
+
+export type Options = typeof options;
+
+export type OptionName = keyof Options;
+
+export type OptionType<T extends OptionName> = OptionTypeFromString<
+    Options[T]["type"]
+>;
+
+export type OptionValues = {
+    [key in OptionName]: OptionType<key>;
+};
+
+export const getOption = async <T extends OptionName>(
+    name: T,
+): Promise<OptionType<T>> =>
+    (await chrome.storage.local.get(name))[name] ?? options[name].default;
+
+export const getOptions = async (): Promise<OptionValues> => {
+    const values = (await chrome.storage.local.get(null)) as OptionValues;
+
+    return map(options, (option, key) => values[key] ?? option.default);
+};
 
 export default options;
