+++ src/content-scripts/modules/base/once.ts
@@ -1,0 +1,11 @@
+import type { Module } from "..";
+
+export const onceModule = (f: () => unknown): Module => {
+    let done = false;
+
+    return () => {
+        if (done) return;
+        done = true;
+        return f();
+    };
+};
