--- src/content-scripts/modules/options/autoLogin.ts
@@ -1,5 +1,0 @@
-import { rememberLogin } from "../login";
-
-export default (value: boolean) => {
-    if (value) rememberLogin();
-};
