--- src/content-scripts/modules/utilities/removeElement.ts
@@ -1,5 +1,0 @@
-// Utility function to remove DOM element
-export default function removeElement(query: string) {
-    const element = document.querySelector(query);
-    element && element.remove();
-}
