+++ src/content-scripts/options/shortcuts/index.css
@@ -1,0 +1,3 @@
+#caixa-atalhos {
+    display: none !important;
+}
