+++ src/content-scripts/options/shortcuts/index.ts
@@ -1,0 +1,5 @@
+import { toggleStyleOption } from "../base/style";
+
+export default toggleStyleOption(
+    "src/content-scripts/options/shortcuts/index.css",
+);
